requirements.txt
Django~=2.0.0
channels~=2.0,>=2.0.2
channels_redis~=2.0
pycrypto==2.6.1
